Update for new Alignak daemons implementation Fix broken test with new Alignak version Update PyPI and Bintray (deb/rpm) packaging Update packaging and RPM deployment
