+++ .bintray-deb.json
@@ -1,0 +1,31 @@
+{
+  "package": {
+    "name": "sed_package_name",
+    "repo": "sed_version_repo",
+    "subject": "alignak",
+    "desc": "Alignak backend importation script for Nagios configuration files",
+    "vcs_url": "https://github.com/Alignak-monitoring-contrib/alignak-backend-import.git",
+    "website_url": "http://alignak.net",
+    "issue_tracker_url": "https://github.com/Alignak-monitoring-contrib/alignak-backend-import/issues",
+    "labels": ["alignak", "backend", "nagios", "monitoring", "REST", "MongoDB"],
+    "licenses": ["AGPL-V3"]
+  },
+  "version": {
+    "name": "sed_version_name",
+    "desc": "sed_version_desc",
+    "released": "sed_version_released"
+  },
+  "files": [
+    {
+      "includePattern": "./dist/(.*\\.deb)",
+      "uploadPattern": "$1",
+      "matrixParams": {
+        "override": "1",
+        "deb_distribution": "jessie,Stretch,Trusty,xenial,Yakkety,Zesty,Artful",
+        "deb_component": "main",
+        "deb_architecture": "all"
+      }
+    }
+  ],
+  "publish": true
+}
