+++ .bintray-rpm.json
@@ -1,0 +1,28 @@
+{
+  "package": {
+    "name": "sed_package_name",
+    "repo": "sed_version_repo",
+    "subject": "alignak",
+    "desc": "Alignak backend importation script",
+    "vcs_url": "https://github.com/Alignak-monitoring-contrib/alignak-backend-import.git",
+    "website_url": "http://alignak.net",
+    "issue_tracker_url": "https://github.com/Alignak-monitoring-contrib/alignak-backend-import/issues",
+    "labels": ["alignak", "backend", "nagios", "monitoring", "REST", "MongoDB"],
+    "licenses": ["AGPL-V3"]
+  },
+  "version": {
+    "name": "sed_version_name",
+    "desc": "sed_version_desc",
+    "released": "sed_version_released"
+  },
+  "files": [
+    {
+      "includePattern": "./dist/(.*\\.rpm)",
+      "uploadPattern": "$1",
+      "matrixParams": {
+        "override": 1
+      }
+    }
+  ],
+  "publish": true
+}
