+++ .travis.yml
@@ -1,28 +1,103 @@
 language: python
 sudo: true
 python:
-  - "2.7"
-  # - "3.4"
+  - '3.6'
+  - '3.5'
+  - '3.4'
+  - '2.7'
 
 services:
   - mongodb
 
-# command to install dependencies
-# some are only used for travis/coveralls so we are installing them here only
 install:
-  - bash setup_test.sh
-# command to run tests
-# notice: the nose-cov is used because it is compatible with --processes, but produce a .coverage by process
-# so we must combine them in the end
-script:
-  # so to help eventual debug: knowing what exact versions are in use can be rather useful.
+  # Install all Python requirements (application and tests)
+  - ./test/setup_test.sh
+  # To help eventual debug: knowing which exact versions are in use can be rather useful.
   - pip freeze
-  # Code static analysis
-  - pycodestyle --max-line-length=100 --exclude='*.pyc, *.cfg, *.log' --ignore='E402' alignak_backend_import/*
-  - pylint --rcfile=.pylintrc alignak_backend_import/*
-  - pep257 --select=D300 alignak_backend_import
+
+script:
   # Code dynamic analysis
   - cd test
-  # No coverage tests
+  # No code coverage tests
   - pytest test_*.py
   - cd ..
+
+jobs:
+  include:
+    - stage: Coding standard
+      python: 3.6
+      script:
+        # Static code analysis
+        - pycodestyle --max-line-length=100 --exclude='*.pyc, *.cfg, *.log' --ignore='E402' alignak_backend_import/*
+        - pylint --rcfile=.pylintrc alignak_backend_import/*
+        - pep257 --select=D300 alignak_backend_import
+    - stage: Pypi deployment
+      python: 3.6
+      script: echo "Deploying to PyPi..."
+      if: (tag IS present)
+      deploy:
+      - provider: pypi
+        distributions: "sdist bdist_wheel"
+        skip_existing: true
+        on:
+          tags: true
+        user: alignak
+        password:
+          secure: "K73Gxkli34xbWcAn3Q76Cmbg0isBG6REwvyiOOBEGwACRK/B3+lhpXFIGxyjPC/1LuLtGCtyJ8VCWNDvnKyITFrm4iNwmc1rTNLqh+PI7kZmENeX1OkIwDN7g/rcVzVIIu3LAGboJ8U2UeGS3aJChncq7Wg6J9ZkMnXxAC/b9flUp9jgnAq9wLdGRQJMVcvq/3kwpNZ9EwgUU5VSV+2Zuwj6PF6pLOX4FbYddxEWO4mcHWuT+A4SKSDLF0mfCkBo91SrLwN/dV4amk4y3nPTo78CdSsivuexu/goIWgJgP3w3OAUCXXW9SnsgTQzUgkXtmBNYufeC0OSDAYqODXPNuh9alQ0Qhxmd4Bstxa+phSaDGP9SL9Mh2VQ4af7nksPLyc3bPqhIZgOnoOu2E2N2yO826vFeC2gdtddS9sZrMAwCSoHS2gvV13CmruKweWFmoqxBJAoRHUHsmu58ndg0l24QV361EXD0pJxt+46SlYHz1QdzqDtxeBXqsAFMC9Dt/0SVeBaCd0skuUao+y5dpOx/gHKSMYiEHEs+ybIq+swXAKNcxHvFNrae6UpMnEHBgjlQrasqBECXn83nVxxU8aPgrnWfUweLowhzD6TLmgrrTsFI1qMX3Wpj0em7dd0FwAs7ntrGKQoRZTLauLe7EfeTADkqYbWZe60tIvEwvw="
+    - stage: Debian packaging
+      python: 3.6
+      install:
+        # Install packaging requirements
+        - echo "Installing packaging requirements..."
+      script:
+        - echo "Deploying to BinTray packages..."
+        # Package Alignak for debian
+        - ./package.sh $TRAVIS_BRANCH $TRAVIS_PYTHON_VERSION deb
+      deploy:
+      - provider: bintray
+        skip_cleanup: true
+        file: dist/.bintray-deb.json
+        on:
+          all_branches: true
+        user: mohierf
+        key:
+          secure: "ARqItCMxxSnjCeo8eiLohcf0EWbmhdHAXeozYzVuelijrr4v4yW9VfwYB72kgD8vrM/4Yx1d2rnzJinPXzZjsYEsj52aVzYcJ17WkK0W9eYZSnxkD7YRBmpr5hiL34M3FRoIYsoLd8OLrZuFKbRWy6zy2kQqOQVHpXpunylGSMtMig4xkGQx+1fAlk9kzMiLZwmWeiQfO/NWWfKNxl4M37LqoT1xzYA3BQjgWgHv3iB16eCZEE78Tn7DajDmgNkmFlHy2i4oel9pD08ZiPTX2TAyS1pP4tdmHe4EXTL2O6OAKLyWyXsz+xqft0G7l0F+HwB835XtE3GlgMUnmkFm7iTwDnP9D/WlfmkP7F5w9MNLIOpwwcO9JLd2zES3A2Yf9AxMtOb/y9XOisZ9/NaFeAigYmAHvS0Xlqoq4XRrWFgqNoyK3r6naMtWqgGsbRJdTcixwWodTSs2NLOuwCVQu+m0SJRmA+gVfCrA3n4D3xf/tQR13bOgr8m5Dw7N7Wg1b4sBXCjjL/i+iPpl8CTb+I4qMvF2wevmtffx8nvfTYwUAvCy3OWfHfX74CetGzX4IgEarR+gi70UiUkhJI6ajbW20/PapY3H/VLhHyye6jnXf1k9VnQRJhF1+EZUhkj5NxU9yjEM8ARlEKQ8u0rcJ9lL4UmpjZWabBnqo12rhNs="
+    - stage: RPM packaging
+      python: 2.7
+      install:
+        # Install packaging requirements
+        - echo "Installing packaging requirements..."
+        - sudo apt-get install -y rpm
+      script:
+        - echo "Deploying to BinTray packages..."
+        # Package Alignak-backend for deb/rpm
+        - ./package.sh $TRAVIS_BRANCH $TRAVIS_PYTHON_VERSION rpm
+      deploy:
+      - provider: bintray
+        skip_cleanup: true
+        file: dist/.bintray-rpm.json
+        on:
+          all_branches: true
+        user: mohierf
+        key:
+          secure: "ARqItCMxxSnjCeo8eiLohcf0EWbmhdHAXeozYzVuelijrr4v4yW9VfwYB72kgD8vrM/4Yx1d2rnzJinPXzZjsYEsj52aVzYcJ17WkK0W9eYZSnxkD7YRBmpr5hiL34M3FRoIYsoLd8OLrZuFKbRWy6zy2kQqOQVHpXpunylGSMtMig4xkGQx+1fAlk9kzMiLZwmWeiQfO/NWWfKNxl4M37LqoT1xzYA3BQjgWgHv3iB16eCZEE78Tn7DajDmgNkmFlHy2i4oel9pD08ZiPTX2TAyS1pP4tdmHe4EXTL2O6OAKLyWyXsz+xqft0G7l0F+HwB835XtE3GlgMUnmkFm7iTwDnP9D/WlfmkP7F5w9MNLIOpwwcO9JLd2zES3A2Yf9AxMtOb/y9XOisZ9/NaFeAigYmAHvS0Xlqoq4XRrWFgqNoyK3r6naMtWqgGsbRJdTcixwWodTSs2NLOuwCVQu+m0SJRmA+gVfCrA3n4D3xf/tQR13bOgr8m5Dw7N7Wg1b4sBXCjjL/i+iPpl8CTb+I4qMvF2wevmtffx8nvfTYwUAvCy3OWfHfX74CetGzX4IgEarR+gi70UiUkhJI6ajbW20/PapY3H/VLhHyye6jnXf1k9VnQRJhF1+EZUhkj5NxU9yjEM8ARlEKQ8u0rcJ9lL4UmpjZWabBnqo12rhNs="
+    - stage: FreeBSD packaging
+      python: 3.6
+      install:
+        # Install packaging requirements
+        - echo "Installing packaging requirements..."
+        - sudo pip install virtualenv virtualenv-tools
+        - sudo pip install --upgrade distribute
+      script:
+        - echo "Deploying to BinTray packages..."
+        # Package Alignak-backend for deb/rpm
+        - ./package.sh $TRAVIS_BRANCH $TRAVIS_PYTHON_VERSION freebsd
+      deploy:
+      - provider: bintray
+        skip_cleanup: true
+        file: dist/.bintray-freebsd.json
+        on:
+          all_branches: true
+        user: mohierf
+        key:
+          secure: "ARqItCMxxSnjCeo8eiLohcf0EWbmhdHAXeozYzVuelijrr4v4yW9VfwYB72kgD8vrM/4Yx1d2rnzJinPXzZjsYEsj52aVzYcJ17WkK0W9eYZSnxkD7YRBmpr5hiL34M3FRoIYsoLd8OLrZuFKbRWy6zy2kQqOQVHpXpunylGSMtMig4xkGQx+1fAlk9kzMiLZwmWeiQfO/NWWfKNxl4M37LqoT1xzYA3BQjgWgHv3iB16eCZEE78Tn7DajDmgNkmFlHy2i4oel9pD08ZiPTX2TAyS1pP4tdmHe4EXTL2O6OAKLyWyXsz+xqft0G7l0F+HwB835XtE3GlgMUnmkFm7iTwDnP9D/WlfmkP7F5w9MNLIOpwwcO9JLd2zES3A2Yf9AxMtOb/y9XOisZ9/NaFeAigYmAHvS0Xlqoq4XRrWFgqNoyK3r6naMtWqgGsbRJdTcixwWodTSs2NLOuwCVQu+m0SJRmA+gVfCrA3n4D3xf/tQR13bOgr8m5Dw7N7Wg1b4sBXCjjL/i+iPpl8CTb+I4qMvF2wevmtffx8nvfTYwUAvCy3OWfHfX74CetGzX4IgEarR+gi70UiUkhJI6ajbW20/PapY3H/VLhHyye6jnXf1k9VnQRJhF1+EZUhkj5NxU9yjEM8ARlEKQ8u0rcJ9lL4UmpjZWabBnqo12rhNs="
