+++ setup.py
@@ -1,7 +1,7 @@
 ﻿#!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-# Copyright (c) 2015:
+# Copyright (c) 2015-2018:
 #   Frederic Mohier, frederic.mohier@gmail.com
 #
 # This file is part of (alignak_backend_import).
@@ -30,45 +30,25 @@ try:
 except:
     sys.exit("Error: missing python-setuptools library")
 
+long_description = "Python Alignak backend importation tool"
+try:
+    with open('README.rst') as f:
+        long_description = f.read()
+except IOError:
+    pass
+
 try:
     python_version = sys.version_info
 except:
     python_version = (1, 5)
 if python_version < (2, 7):
     sys.exit("This application requires a minimum Python 2.7.x, sorry!")
-elif python_version >= (3,):
-    sys.exit("This application is not yet compatible with Python 3.x, sorry!")
 
-from alignak_backend_import import __application__, __version__, __author__, __author_email__
-from alignak_backend_import import  __copyright__, __releasenotes__, __license__, __doc_url__
+from alignak_backend_import import __version__, __author__, __author_email__, __doc_url__
+from alignak_backend_import import  __copyright__, __classifiers__, __license__, __git_url__
 from alignak_backend_import import __name__ as __pkg_name__
 
 package = import_module('alignak_backend_import')
-
-# Define paths
-if 'linux' in sys.platform or 'sunos5' in sys.platform:
-    paths = {
-        'bin':     "/usr/bin",
-        'var':     "/var/lib/alignak_backend_import/",
-        'share':   "/var/lib/alignak_backend_import/share",
-        'etc':     "/etc/alignak_backend_import",
-        'run':     "/var/run/alignak_backend_import",
-        'log':     "/var/log/alignak_backend_import",
-        'libexec': "/var/lib/alignak_backend_import/libexec",
-    }
-elif 'bsd' in sys.platform or 'dragonfly' in sys.platform:
-    paths = {
-        'bin':     "/usr/local/bin",
-        'var':     "/usr/local/libexec/alignak_backend_import",
-        'share':   "/usr/local/share/alignak_backend_import",
-        'etc':     "/usr/local/etc/alignak_backend_import",
-        'run':     "/var/run/alignak_backend_import",
-        'log':     "/var/log/alignak_backend_import",
-        'libexec': "/usr/local/libexec/alignak_backend_import/plugins",
-    }
-else:
-    print "Unsupported platform, sorry!"
-    exit(1)
 
 setup(
     name=__pkg_name__,
@@ -79,18 +59,34 @@ setup(
     # metadata for upload to PyPI
     author=__author__,
     author_email=__author_email__,
-    keywords="alignak REST backend import tool",
-    url="https://github.com/Alignak-monitoring-contrib/alignak-backend-import",
+    url=__doc_url__,
+    download_url=__git_url__,
+    keywords="alignak monitoring backend import tool",
     description=package.__doc__.strip(),
-    long_description=open('README.rst').read(),
+    long_description=long_description,
+    long_description_content_type='text/x-rst',
 
-    zip_safe=False,
+    project_urls={
+        'Presentation': 'http://alignak.net',
+        'Documentation': 'http://docs.alignak.net/en/latest/',
+        'Source': 'https://github.com/alignak-monitoring-contrib/alignak-backed/',
+        'Tracker': 'https://github.com/alignak-monitoring-contrib/alignak-backend/issues',
+        'Contributions': 'https://github.com/alignak-monitoring-contrib/'
+    },
 
-    packages=find_packages(),
+    # Package data
+    packages=find_packages(exclude=['docs', 'test']),
     include_package_data=True,
 
-    install_requires=[
-        'docopt', 'future', 'alignak-backend-client'
+    # Unzip Egg
+    zip_safe=False,
+    platforms='any',
+
+    # Dependencies...
+    install_requires=['future', 'requests', 'alignak-backend-client'],
+    dependency_links=[
+        # Use the standard PyPi repository
+        "https://pypi.python.org/simple/",
     ],
 
     entry_points={
@@ -99,15 +95,5 @@ setup(
         ],
     },
 
-    classifiers = [
-        'Development Status :: 5 - Production/Stable',
-        'Environment :: Console',
-        'Intended Audience :: Developers',
-        'Intended Audience :: System Administrators',
-        'License :: OSI Approved :: GNU General Public License v3 (GPLv3)',
-        'Natural Language :: English',
-        'Programming Language :: Python',
-        'Topic :: System :: Monitoring',
-        'Topic :: System :: Systems Administration'
-    ]
+    classifiers = __classifiers__
 )
