+++ test/alignak-logger.json
@@ -1,0 +1,72 @@
+{
+    "version": 1,
+    "disable_existing_loggers": false,
+    "formatters": {
+        "alignak": {
+            "format": "[%(asctime)s] %(levelname)s: [%(daemon)s.%(name)s] %(message)s"
+        ,"datefmt": "%Y-%m-%d %H:%M:%S"
+        },
+        "monitoring-log": {
+            "format": "[%(asctime)s] %(levelname)s: %(message)s"
+        ,"datefmt": "%Y-%m-%d %H:%M:%S"
+        }
+    },
+
+    "handlers": {
+        "unit_tests": {
+            "class": "alignak.log.CollectorHandler",
+            "level": "DEBUG",
+            "formatter": "alignak"
+        },
+        "console": {
+            "class": "logging.StreamHandler",
+            "level": "DEBUG",
+            "formatter": "alignak",
+            "stream": "ext://sys.stdout"
+        },
+        "color_console": {
+            "class": "alignak.log.ColorStreamHandler",
+            "level": "DEBUG",
+            "formatter": "alignak",
+            "stream": "ext://sys.stdout"
+        },
+        "daemons": {
+            "class": "logging.handlers.TimedRotatingFileHandler",
+            "level": "DEBUG",
+            "formatter": "alignak",
+            "filename": "%(logdir)s/%(daemon)s.log",
+            "when": "midnight",
+            "interval": 1,
+            "backupCount": 7,
+            "encoding": "utf8"
+        },
+        "monitoring_logs": {
+            "class": "logging.handlers.TimedRotatingFileHandler",
+            "level": "INFO",
+            "formatter": "monitoring-log",
+            "filename": "%(logdir)s/monitoring-logs.log",
+            "when": "midnight",
+            "interval": 1,
+            "backupCount": 365,
+            "encoding": "utf8"
+        }
+    },
+
+    "loggers": {
+        "alignak": {
+            "level": "INFO",
+            "handlers": ["unit_tests", "color_console", "daemons"],
+            "propagate": "no"
+        },
+        "monitoring-log": {
+            "level": "DEBUG",
+            "handlers": ["console", "monitoring_logs"],
+            "propagate": "no"
+        }
+    },
+
+    "root": {
+        "level": "ERROR",
+        "handlers": []
+    }
+}
