+++ test/alignak_cfg_files/hostgroups_links_hostgroup.cfg
@@ -37,7 +37,8 @@ define hostgroup{
 }
 
 define timeperiod{
-    timeperiod_name	24x7
+    timeperiod_name			24x7
+    alias							Always
 }
 
 define command{
