+++ test/alignak_cfg_files/hosts_custom_variables.cfg
@@ -4,7 +4,7 @@ define host{
        max_check_attempts                 6
 
 # Custom in the host template
-   _LOC_LAT             45
+   #_LOC_LAT             45
 
        register                           0
 }
