+++ test/alignak_cfg_files/hosts_links_hostgroup.cfg
@@ -36,7 +36,8 @@ define hostgroup{
 }
 
 define timeperiod{
-	timeperiod_name	24x7
+	timeperiod_name			24x7
+	alias							Always
 }
 
 define command{
