+++ test/alignak_cfg_files/hosts_links_parent.cfg
@@ -45,7 +45,8 @@ define host{
 }
 
 define timeperiod{
-	timeperiod_name	24x7
+   timeperiod_name			24x7
+   alias							Always
 }
 
 define command{
