+++ test/alignak_cfg_files/hosts_services_templates.cfg
@@ -12,7 +12,8 @@ define host {
 
 # Default host template
 define timeperiod{
-    timeperiod_name	24x7
+    timeperiod_name			24x7
+    alias							Always
 }
 define host{
 	name				    generic-host
