+++ test/alignak_cfg_files/services_link_servicegroups.cfg
@@ -75,7 +75,8 @@ define servicegroup{
 }
 
 define timeperiod{
-	timeperiod_name	24x7
+   timeperiod_name			24x7
+   alias							Always
 }
 
 define command{
