+++ test/shinken_cfg_files/default/timeperiods/24x7.cfg
@@ -1,6 +1,6 @@
 define timeperiod{
 	timeperiod_name			24x7
-	alias				Always
+	alias							Always
 	sunday				00:00-24:00
 	monday				00:00-24:00
 	tuesday				00:00-24:00
