+++ test/test_cfg_to_backend.py
@@ -20,6 +20,8 @@ class TestCfgToBackend(unittest2.TestCase):
         os.environ['ALIGNAK_BACKEND_MONGO_DBNAME'] = 'alignak-backend-import-test'
         os.environ['ALIGNAK_BACKEND_CONFIGURATION_FILE'] = './cfg/settings/settings.json'
 
+        cls.maxDiff = None
+
         # Delete used mongo DBs
         exit_code = subprocess.call(
             shlex.split('mongo %s --eval "db.dropDatabase()"'
@@ -32,17 +34,12 @@ class TestCfgToBackend(unittest2.TestCase):
         print("Current test directory: %s" % test_dir)
 
         print("Starting Alignak backend...")
-        fnull = open(os.devnull, 'w')
-        cls.backend_process = subprocess.Popen(shlex.split('alignak-backend'))
-        print("Started as %s" % cls.backend_process.pid)
-
-        # cls.p = subprocess.Popen(['uwsgi', '--plugin', 'python', '-w', 'alignak_backend.app:app',
-        #                           '--socket', '0.0.0.0:5000',
-        #                           '--protocol=http', '--enable-threads', '--pidfile',
-        #                           '/tmp/uwsgi.pid'])
-        # time.sleep(3)
-
-        time.sleep(3)
+        cls.p = subprocess.Popen(['uwsgi', '--plugin', 'python', '-w', 'alignak_backend.app:app',
+                                  '--socket', '0.0.0.0:5000',
+                                  '--protocol=http', '--enable-threads', '--pidfile',
+                                  '/tmp/uwsgi.pid'])
+        time.sleep(1)
+        print("Started as %s" % cls.p.pid)
 
         cls.backend = Backend('http://127.0.0.1:5000')
         cls.backend.login("admin", "admin")
@@ -61,7 +58,6 @@ class TestCfgToBackend(unittest2.TestCase):
         """
         print("Stopping Alignak backend...")
         subprocess.call(['uwsgi', '--stop', '/tmp/uwsgi.pid'])
-        time.sleep(2)
         print("Stopped")
 
     @classmethod
@@ -92,10 +88,11 @@ class TestCfgToBackend(unittest2.TestCase):
                    u"notes": u"",
                    u'_sub_realm': True,
                    u"alias": u"Normal Work Hours",
-                   u"dateranges": [{u'monday': u'09:00-17:00'}, {u'tuesday': u'09:00-17:00'},
-                                   {u'friday': u'09:00-12:00,14:00-16:00'},
+                   u"dateranges": [{u'monday': u'09:00-17:00'},
+                                   {u'tuesday': u'09:00-17:00'},
                                    {u'wednesday': u'09:00-17:00'},
-                                   {u'thursday': u'09:00-17:00'}],
+                                   {u'thursday': u'09:00-17:00'},
+                                   {u'friday': u'09:00-12:00,14:00-16:00'}],
                    u"exclude": [], u"is_active": False, u"imported_from": u"alignak-backend-import"
                    }
             del comm['_links']
@@ -105,7 +102,13 @@ class TestCfgToBackend(unittest2.TestCase):
             del comm['_updated']
             del comm['_realm']
             del comm['schema_version']
+            dr1 = comm.pop('dateranges')
+            dr2 = ref.pop('dateranges')
             self.assertEqual(comm, ref)
+            # Dateranges are the same ?
+            for dr in dr1:
+                assert dr in dr2
+            # assert not [k for k in dr1 if dr1[k] != dr2[k]]
         self.assertTrue(found)
 
     def test_timeperiod_duplicates(self):
@@ -140,10 +143,12 @@ class TestCfgToBackend(unittest2.TestCase):
                    u"notes": u"",
                    u'_sub_realm': True,
                    u"alias": u"Normal Work Hours",
-                   u"dateranges": [{u'monday': u'09:00-18:00'}, {u'tuesday': u'09:00-18:00'},
-                                   {u'friday': u'09:00-12:00,14:00-16:00'},
+                   u"dateranges": [{u'monday': u'09:00-18:00'},
+                                   {u'tuesday': u'09:00-18:00'},
+                                   {u'thursday': u'09:00-18:00'},
                                    {u'wednesday': u'09:00-18:00'},
-                                   {u'thursday': u'09:00-18:00'}],
+                                   {u'friday': u'09:00-12:00,14:00-16:00'}
+                                   ],
                    u"exclude": [], u"is_active": False, u"imported_from": u"alignak-backend-import"
                    }
             del comm['_links']
@@ -153,7 +158,13 @@ class TestCfgToBackend(unittest2.TestCase):
             del comm['_updated']
             del comm['_realm']
             del comm['schema_version']
+            dr1 = comm.pop('dateranges')
+            dr2 = ref.pop('dateranges')
             self.assertEqual(comm, ref)
+            # Dateranges are the same ?
+            for dr in dr1:
+                assert dr in dr2
+            # assert not [k for k in dr1 if dr1[k] != dr2[k]]
         self.assertTrue(found)
 
     def test_timeperiod_update(self):
@@ -180,10 +191,12 @@ class TestCfgToBackend(unittest2.TestCase):
                    u"notes": u"",
                    u'_sub_realm': True,
                    u"alias": u"Normal Work Hours",
-                   u"dateranges": [{u'monday': u'09:00-18:00'}, {u'tuesday': u'09:00-18:00'},
-                                   {u'friday': u'09:00-12:00,14:00-16:00'},
+                   u"dateranges": [{u'monday': u'09:00-18:00'},
+                                   {u'tuesday': u'09:00-18:00'},
                                    {u'wednesday': u'09:00-18:00'},
-                                   {u'thursday': u'09:00-18:00'}],
+                                   {u'thursday': u'09:00-18:00'},
+                                   {u'friday': u'09:00-12:00,14:00-16:00'}
+                                   ],
                    u"exclude": [], u"is_active": False, u"imported_from": u"alignak-backend-import"
                    }
             del comm['_links']
@@ -193,7 +206,13 @@ class TestCfgToBackend(unittest2.TestCase):
             del comm['_updated']
             del comm['_realm']
             del comm['schema_version']
+            dr1 = comm.pop('dateranges')
+            dr2 = ref.pop('dateranges')
             self.assertEqual(comm, ref)
+            # Dateranges are the same ?
+            for dr in dr1:
+                assert dr in dr2
+            # assert not [k for k in dr1 if dr1[k] != dr2[k]]
         self.assertTrue(found)
 
         # Update an existing element
@@ -231,7 +250,13 @@ class TestCfgToBackend(unittest2.TestCase):
             del comm['_updated']
             del comm['_realm']
             del comm['schema_version']
+            dr1 = comm.pop('dateranges')
+            dr2 = ref.pop('dateranges')
             self.assertEqual(comm, ref)
+            # Dateranges are the same ?
+            for dr in dr1:
+                assert dr in dr2
+            # assert not [k for k in dr1 if dr1[k] != dr2[k]]
         self.assertTrue(found)
 
     def test_timeperiod_complex(self):
@@ -465,7 +490,7 @@ class TestCfgToBackend(unittest2.TestCase):
         self.assertEqual(exit_code, 0)
 
         c = self.backend.get('command')
-        self.assertEqual(len(c['_items']), 4)
+        self.assertEqual(len(c['_items']), 6)
         command_id = ''
         ev_handler_id = ''
         for co in c['_items']:
@@ -496,10 +521,11 @@ class TestCfgToBackend(unittest2.TestCase):
         self.assertEqual(exit_code, 0)
 
         c = self.backend.get('command')
-        self.assertEqual(len(c['_items']), 4)
+        self.assertEqual(len(c['_items']), 6)
         command_id = ''
         ev_handler_id = ''
         for co in c['_items']:
+            print("Command: %s" % co)
             if co['name'] == 'check_tcp':
                 command_id = co['_id']
             if co['name'] == 'my_host_event_handler':
@@ -535,539 +561,5 @@ class TestCfgToBackend(unittest2.TestCase):
         for host in hosts:
             print("- %s, customs: %s" % (host['name'], host['customs']))
             self.assertEqual(host['customs'], {
-                u'_DISPLAY_NAME': u'srv01', u'_LOC_LAT': u'45.054700', u'_LOC_LNG': u'5.080856'
+                u'_LOC_LAT': u'45.054700', u'_LOC_LNG': u'5.080856'
             })
-
-
-class TestContactsNW(unittest2.TestCase):
-    @classmethod
-    def setUpClass(cls):
-        # Set test mode for applications backend
-        os.environ['TEST_ALIGNAK_BACKEND'] = '1'
-        os.environ['ALIGNAK_BACKEND_MONGO_DBNAME'] = 'alignak-backend-import-test'
-
-        # Delete used mongo DBs
-        exit_code = subprocess.call(
-            shlex.split('mongo %s --eval "db.dropDatabase()"'
-                        % os.environ['ALIGNAK_BACKEND_MONGO_DBNAME'])
-        )
-        assert exit_code == 0
-        time.sleep(1)
-
-        test_dir = os.path.dirname(os.path.realpath(__file__))
-        print("Current test directory: %s" % test_dir)
-
-        print("Starting Alignak backend...")
-        fnull = open(os.devnull, 'w')
-        cls.backend_process = subprocess.Popen(shlex.split('alignak-backend'))
-        print("Started as %s" % cls.backend_process.pid)
-
-        time.sleep(3)
-
-        cls.backend = Backend('http://127.0.0.1:5000')
-        cls.backend.login("admin", "admin")
-
-        cls.backend.delete("host", {})
-        cls.backend.delete("service", {})
-        cls.backend.delete("command", {})
-        cls.backend.delete("livesynthesis", {})
-
-    @classmethod
-    def tearDownClass(cls):
-        """
-        Stop alignak backend
-
-        :return: None
-        """
-        print("Stopping Alignak backend...")
-        cls.backend_process.kill()
-        print("Stopped")
-
-    def test_user_notification_ways(self):
-        q = subprocess.Popen(['../alignak_backend_import/cfg_to_backend.py', '--delete', '--quiet',
-                              'alignak_cfg_files/users.cfg'])
-        (_, _) = q.communicate()
-        exit_code = q.wait()
-        self.assertEqual(exit_code, 0)
-
-        result = self.backend.get('command')
-        cmds = result['_items']
-        for cmd in cmds:
-            print(cmd['_id'], cmd['name'])
-            if cmd['name'] == 'notify-host-by-email':
-                cmd_nh1 = cmd['_id']
-            if cmd['name'] == 'notify-host-by-email-2':
-                cmd_nh2 = cmd['_id']
-            if cmd['name'] == 'notify-service-by-email':
-                cmd_ns = cmd['_id']
-        self.assertGreaterEqual(len(cmds), 3)
-        self.assertIsNotNone(cmd_nh1)
-        self.assertIsNotNone(cmd_nh2)
-        self.assertIsNotNone(cmd_ns)
-
-        result = self.backend.get_all('user', params={'where': json.dumps({'_is_template': False})})
-        users = result['_items']
-        self.assertEqual(len(users), 5)
-
-        print("Found users: ")
-        for user in users:
-            print("-", user['name'])
-            if user['name'] == 'admin':
-                self.assertEqual(user['is_admin'], True)
-                self.assertEqual(user['back_role_super_admin'], True)
-            else:
-                self.assertEqual(user['back_role_super_admin'], False)
-                self.assertTrue(user['host_notifications_enabled'])
-                self.assertTrue(user['service_notifications_enabled'])
-                self.assertEqual(user['host_notification_commands'], [cmd_nh1, cmd_nh2])
-                self.assertEqual(user['service_notification_commands'], [cmd_ns])
-
-
-class TestContacts(unittest2.TestCase):
-    @classmethod
-    def setUpClass(cls):
-        # Set test mode for applications backend
-        os.environ['TEST_ALIGNAK_BACKEND'] = '1'
-        os.environ['ALIGNAK_BACKEND_MONGO_DBNAME'] = 'alignak-backend-import-test'
-
-        # Delete used mongo DBs
-        exit_code = subprocess.call(
-            shlex.split('mongo %s --eval "db.dropDatabase()"'
-                        % os.environ['ALIGNAK_BACKEND_MONGO_DBNAME'])
-        )
-        assert exit_code == 0
-        time.sleep(1)
-
-        test_dir = os.path.dirname(os.path.realpath(__file__))
-        print("Current test directory: %s" % test_dir)
-
-        print("Starting Alignak backend...")
-        fnull = open(os.devnull, 'w')
-        cls.backend_process = subprocess.Popen(shlex.split('alignak-backend'))
-        print("Started as %s" % cls.backend_process.pid)
-
-        time.sleep(3)
-
-        cls.backend = Backend('http://127.0.0.1:5000')
-        cls.backend.login("admin", "admin")
-
-        cls.backend.delete("host", {})
-        cls.backend.delete("service", {})
-        cls.backend.delete("command", {})
-        cls.backend.delete("livesynthesis", {})
-
-    @classmethod
-    def tearDownClass(cls):
-        """
-        Stop alignak backend
-
-        :return: None
-        """
-        print("Stopping Alignak backend...")
-        cls.backend_process.kill()
-        print("Stopped")
-
-    def test_user_template(self):
-        """Tests users templates"""
-        # Import and insert the templates into the backend (--model)
-        q = subprocess.Popen(['../alignak_backend_import/cfg_to_backend.py', '--delete', '--quiet',
-                              'alignak_cfg_files/users.cfg'])
-        (_, _) = q.communicate()
-        exit_code = q.wait()
-        self.assertEqual(exit_code, 0)
-
-        result = self.backend.get_all('user')
-        users = result['_items']
-        self.assertEqual(len(users), 6)
-
-        print("Found users: ")
-        self.generic_contact = None
-        for user in users:
-            print("-", user['name'], user)
-            # generic-contact only is a template
-            if user['name'] == 'generic-contact':
-                self.assertEqual(user['_is_template'], True)
-                self.generic_contact = user['_id']
-        self.assertIsNotNone(self.generic_contact)
-
-        print("Found users: ")
-        for user in users:
-            print("-", user['name'], user)
-            if user['name'] == 'generic-contact':
-                continue
-
-            self.assertEqual(user['_is_template'], False)
-            # Except the admin user, all are tagged with the generic-contact
-            if user['name'] != 'admin':
-                self.assertEqual(user['_templates'], [self.generic_contact])
-                # Tags remain empty
-                self.assertEqual(user['tags'], [])
-
-    def test_user_direct_notification(self):
-        """Test user direct notifications"""
-        q = subprocess.Popen(['../alignak_backend_import/cfg_to_backend.py', '--delete', '--quiet',
-                              'alignak_cfg_files/user_admin.cfg'])
-        (_, _) = q.communicate()
-        exit_code = q.wait()
-        self.assertEqual(exit_code, 0)
-
-        result = self.backend.get('command')
-        cmds = result['_items']
-        for cmd in cmds:
-            print("Command: %s" % cmd)
-            if cmd['name'] == 'notify-host-by-email':
-                cmd_nh1 = cmd['_id']
-            if cmd['name'] == 'notify-service-by-email':
-                cmd_ns = cmd['_id']
-        self.assertGreaterEqual(len(cmds), 2)
-        self.assertIsNotNone(cmd_nh1)
-        self.assertIsNotNone(cmd_ns)
-
-        result = self.backend.get_all('user')
-        users = result['_items']
-        self.assertEqual(len(users), 2)
-
-        print("Found users: ")
-        for user in users:
-            print("-", user['name'], user)
-            if user['name'] == 'admin':
-                self.assertEqual(user['is_admin'], True)
-                self.assertEqual(user['back_role_super_admin'], True)
-                self.assertEqual(user['can_update_livestate'], True)
-            else:
-                self.assertEqual(user['is_admin'], True)
-                self.assertEqual(user['back_role_super_admin'], False)
-                self.assertEqual(user['can_update_livestate'], False)
-                self.assertEqual(user['host_notification_commands'], [cmd_nh1])
-                self.assertEqual(user['service_notification_commands'], [cmd_ns])
-
-
-class TestHosts(unittest2.TestCase):
-    @classmethod
-    def setUpClass(cls):
-        # Set test mode for applications backend
-        os.environ['TEST_ALIGNAK_BACKEND'] = '1'
-        os.environ['ALIGNAK_BACKEND_MONGO_DBNAME'] = 'alignak-backend-import-test'
-
-        # Delete used mongo DBs
-        exit_code = subprocess.call(
-            shlex.split('mongo %s --eval "db.dropDatabase()"'
-                        % os.environ['ALIGNAK_BACKEND_MONGO_DBNAME'])
-        )
-        assert exit_code == 0
-        time.sleep(1)
-
-        test_dir = os.path.dirname(os.path.realpath(__file__))
-        print("Current test directory: %s" % test_dir)
-
-        print("Starting Alignak backend...")
-        fnull = open(os.devnull, 'w')
-        cls.backend_process = subprocess.Popen(shlex.split('alignak-backend'))
-        print("Started as %s" % cls.backend_process.pid)
-
-        time.sleep(3)
-
-        cls.backend = Backend('http://127.0.0.1:5000')
-        cls.backend.login("admin", "admin")
-
-        cls.backend.delete("host", {})
-        cls.backend.delete("service", {})
-        cls.backend.delete("command", {})
-        cls.backend.delete("livesynthesis", {})
-
-    @classmethod
-    def tearDownClass(cls):
-        """
-        Stop alignak backend
-
-        :return: None
-        """
-        print("Stopping Alignak backend...")
-        cls.backend_process.kill()
-        print("Stopped")
-
-    def test_hosts(self):
-
-        q = subprocess.Popen(['../alignak_backend_import/cfg_to_backend.py', '--delete', '--quiet',
-                              'alignak_cfg_files/hosts.cfg'])
-        (_, _) = q.communicate()
-        exit_code = q.wait()
-        self.assertEqual(exit_code, 0)
-
-        result = self.backend.get('timeperiod')
-        tps = result['_items']
-        for tp in tps:
-            print(tp['_id'], tp['name'])
-            if tp['name'] == '24x7':
-                tp_always = tp['_id']
-            if tp['name'] == 'Never':
-                tp_never = tp['_id']
-            if tp['name'] == 'All time default 24x7':
-                tp_default = tp['_id']
-        self.assertEqual(len(tps), 2)
-
-        result = self.backend.get('host', params={'where': json.dumps({'_is_template': False})})
-        hosts = result['_items']
-        # 3 newly created hosts
-        self.assertEqual(len(hosts), 3)
-        for host in hosts:
-            if host['name'] == '_dummy':
-                continue
-            # Hosts specific fields
-            if host['name'] == 'srv01':
-                self.assertEqual(host['address'], '192.168.1.11')
-            if host['name'] == 'srv02':
-                self.assertEqual(host['address'], '192.168.1.12')
-            if host['name'] == 'srv03':
-                self.assertEqual(host['address'], '192.168.1.13')
-
-            # Host template fields
-            self.assertEqual(host['check_interval'], 4)
-            self.assertEqual(host['max_check_attempts'], 6)
-
-            # Host template fields - must have a valid check period
-            self.assertIn('check_period', host)
-            self.assertEqual(host['check_period'], tp_always)
-
-            # Host template fields - must have a valid notification period
-            self.assertIn('notification_period', host)
-            self.assertEqual(host['notification_period'], tp_always)
-
-            # Host template fields - must have a Never maintenance period
-            self.assertIn('maintenance_period', host)
-            self.assertEqual(host['maintenance_period'], tp_never)
-
-            # Host template fields - must have a Never snapshot period
-            self.assertIn('snapshot_period', host)
-            self.assertEqual(host['snapshot_period'], tp_never)
-
-    def test_hosts_template(self):
-        """Tests hosts templates"""
-        # Import and insert the templates into the backend (--model)
-        q = subprocess.Popen(['../alignak_backend_import/cfg_to_backend.py', '--delete', '--quiet',
-                              'alignak_cfg_files/hosts_2_templates.cfg'])
-        (_, _) = q.communicate()
-        exit_code = q.wait()
-        self.assertEqual(exit_code, 0)
-
-        result = self.backend.get_all('host', params={'where': json.dumps({'_is_template': True})})
-        hosts = result['_items']
-        self.assertEqual(len(hosts), 4)
-
-        print("Found hosts templates: ")
-        self.template1 = None
-        self.template2 = None
-        self.template3 = None
-        for host in hosts:
-            print("-", host['name'], host)
-            # template01 only is a template
-            if host['name'] == 'template01':
-                self.assertEqual(host['_is_template'], True)
-                self.template1 = host['_id']
-            # template02 only is a template
-            if host['name'] == 'template02':
-                self.assertEqual(host['_is_template'], True)
-                self.template2 = host['_id']
-            # template03 only is a template
-            if host['name'] == 'template03':
-                self.assertEqual(host['_is_template'], True)
-                self.template3 = host['_id']
-        self.assertIsNotNone(self.template1)
-        self.assertIsNotNone(self.template2)
-        self.assertIsNotNone(self.template3)
-
-        result = self.backend.get_all('host', params={'where': json.dumps({'_is_template': False})})
-        hosts = result['_items']
-        self.assertEqual(len(hosts), 1)
-
-        print("Found hosts: ")
-        for host in hosts:
-            print("-", host['name'], host)
-            if host['name'] in ['template01', 'template02', '_dummy']:
-                self.assertEqual(host['_is_template'], True)
-                continue
-
-            if host['name'] in ['template03']:
-                self.assertEqual(host['_is_template'], True)
-                self.assertEqual(host['_templates'], [self.template2])
-                # Tags remain empty
-                self.assertEqual(host['tags'], [])
-                continue
-
-            self.assertEqual(host['_is_template'], False)
-            if host['name'] == 'srv01':
-                self.assertEqual(host['_templates'], [self.template1, self.template3])
-                # Tags remain empty
-                self.assertEqual(host['tags'], [])
-
-    def test_host_with_double_template(self):
-
-        q = subprocess.Popen(['../alignak_backend_import/cfg_to_backend.py', '--delete', '--quiet',
-                              'alignak_cfg_files/hosts_2_templates.cfg'])
-        (_, _) = q.communicate()
-        exit_code = q.wait()
-        self.assertEqual(exit_code, 0)
-
-        result = self.backend.get('host', params={'where': json.dumps({'_is_template': False})})
-        # 1 newly created host
-        self.assertEqual(len(result['_items']), 1)
-        for comm in result['_items']:
-            reg_comm = comm.copy()
-
-        self.assertEqual(reg_comm['name'], 'srv01')
-        self.assertEqual(reg_comm['max_check_attempts'], 6)
-        self.assertEqual(reg_comm['check_interval'], 2)
-
-    def test_hosts_dependency(self):
-
-        q = subprocess.Popen(['../alignak_backend_import/cfg_to_backend.py', '--delete', '--quiet',
-                              'alignak_cfg_files/hosts_links_parent.cfg'])
-        (_, _) = q.communicate()
-        exit_code = q.wait()
-        self.assertEqual(exit_code, 0)
-
-        result = self.backend.get('hostdependency')
-        hds = result['_items']
-        for hd in hds:
-            print(hd)
-        self.assertEqual(len(hds), 2)
-
-
-class TestServices(unittest2.TestCase):
-    @classmethod
-    def setUpClass(cls):
-        # Set test mode for applications backend
-        os.environ['TEST_ALIGNAK_BACKEND'] = '1'
-        os.environ['ALIGNAK_BACKEND_MONGO_DBNAME'] = 'alignak-backend-import-test'
-
-        # Delete used mongo DBs
-        exit_code = subprocess.call(
-            shlex.split('mongo %s --eval "db.dropDatabase()"'
-                        % os.environ['ALIGNAK_BACKEND_MONGO_DBNAME'])
-        )
-        assert exit_code == 0
-        time.sleep(1)
-
-        test_dir = os.path.dirname(os.path.realpath(__file__))
-        print("Current test directory: %s" % test_dir)
-
-        print("Starting Alignak backend...")
-        fnull = open(os.devnull, 'w')
-        cls.backend_process = subprocess.Popen(shlex.split('alignak-backend'))
-        print("Started as %s" % cls.backend_process.pid)
-
-        time.sleep(3)
-
-        cls.backend = Backend('http://127.0.0.1:5000')
-        cls.backend.login("admin", "admin")
-
-        cls.backend.delete("host", {})
-        cls.backend.delete("service", {})
-        cls.backend.delete("command", {})
-        cls.backend.delete("livesynthesis", {})
-
-    @classmethod
-    def tearDownClass(cls):
-        """
-        Stop alignak backend
-
-        :return: None
-        """
-        print("Stopping Alignak backend...")
-        cls.backend_process.kill()
-        print("Stopped")
-
-    def test_services_template(self):
-        """Tests hosts templates"""
-        # Import and insert the templates into the backend (--model)
-        q = subprocess.Popen(['../alignak_backend_import/cfg_to_backend.py', '--delete',
-                              'alignak_cfg_files/hosts_services_templates.cfg'])
-        (_, _) = q.communicate()
-        exit_code = q.wait()
-        self.assertEqual(exit_code, 0)
-
-        result = self.backend.get_all('host',
-                                      params={'where': json.dumps({'_is_template': True})})
-        hosts = result['_items']
-        self.assertEqual(len(hosts), 4)
-
-        print("Found host templates: ")
-        host_template = None
-        for host in hosts:
-            print("-", host['name'])
-            self.assertIn(host['name'], ['_dummy', 'generic-host',
-                                         'generic-passive-host', 'windows-passive-host'])
-            if host['name'] == 'windows-passive-host':
-                host_template = host
-        self.assertEqual(len(hosts), 4)
-        self.assertIsNotNone(host_template)
-
-        result = self.backend.get_all('service',
-                                      params={'where': json.dumps({'_is_template': True})})
-        services = result['_items']
-        self.assertEqual(len(services), 6)
-
-        print("Found service templates: ")
-        for service in services:
-            print("-", service['name'])
-            self.assertIn(service['name'], ['windows-passive-service',
-                                            'nsca_uptime',
-                                            'nsca_memory',
-                                            'nsca_cpu',
-                                            'nsca_disk',
-                                            'nsca_services'])
-
-        result = self.backend.get_all('host',
-                                      params={'where': json.dumps({'_is_template': False})})
-        hosts = result['_items']
-        self.assertEqual(len(hosts), 1)
-
-        # Only one host using the template windows-passive-host...
-        print("Found hosts: ")
-        for host in hosts:
-            print("-", host['name'])
-            self.assertEqual(host['_templates'], [host_template['_id']])
-
-        result = self.backend.get_all('service',
-                                      params={'where': json.dumps({'_is_template': False})})
-        services = result['_items']
-        self.assertEqual(len(services), 5)
-
-        # will inherit from 5 services
-        print("Found services: ")
-        for service in services:
-            print("-", service['name'])
-
-    def test_services_hostgroup(self):
-        """Tests services / hostgroups (issue #65)"""
-        # Import and insert the templates into the backend (--model)
-        q = subprocess.Popen(['../alignak_backend_import/cfg_to_backend.py', '--delete',
-                              'alignak_cfg_files/issue65.cfg'])
-        (_, _) = q.communicate()
-        exit_code = q.wait()
-        self.assertEqual(exit_code, 0)
-
-        # Get all hostgroups
-        result = self.backend.get_all('hostgroup')
-        hostgroups = result['_items']
-        self.assertEqual(len(hostgroups), 2)
-        print("Found hostgroups: ")
-        for hostgroup in hostgroups:
-            print("-", hostgroup['name'])
-
-        # Get all hosts
-        result = self.backend.get_all('host', params={'where': json.dumps({'_is_template': False})})
-        hosts = result['_items']
-        self.assertEqual(len(hosts), 3)
-        print("Found hosts: ")
-        for host in hosts:
-            print("-", host['name'])
-
-        # Get all services
-        result = self.backend.get_all('service',
-                                      params={'where': json.dumps({'_is_template': False})})
-        services = result['_items']
-        self.assertEqual(len(services), 3)
-        print("Found services: ")
-        for service in services:
-            print("-", service['name'])
