+++ test/test_import_default.py
@@ -12,6 +12,7 @@ def setup_module(module):
     # Set test mode for applications backend
     os.environ['TEST_ALIGNAK_BACKEND'] = '1'
     os.environ['ALIGNAK_BACKEND_MONGO_DBNAME'] = 'alignak-backend-import-test'
+    os.environ['ALIGNAK_BACKEND_CONFIGURATION_FILE'] = './cfg/settings/settings.json'
 
     # Delete used mongo DBs
     exit_code = subprocess.call(
@@ -25,14 +26,16 @@ def setup_module(module):
     print("Current test directory: %s" % test_dir)
 
     print("Starting Alignak backend...")
-    fnull = open(os.devnull, 'w')
-    subprocess.Popen(shlex.split('alignak-backend'), stdout=fnull)
-    print("Started")
+    subprocess.Popen(['uwsgi', '--plugin', 'python', '-w', 'alignak_backend.app:app',
+                      '--socket', '0.0.0.0:5000',
+                      '--protocol=http', '--enable-threads', '--pidfile',
+                      '/tmp/uwsgi.pid'])
+    time.sleep(1)
 
 
 def teardown_module(module):
     print("Stopping Alignak backend...")
-    subprocess.call(['pkill', 'alignak-backend'])
+    subprocess.call(['uwsgi', '--stop', '/tmp/uwsgi.pid'])
     print("Stopped")
 
 
@@ -46,9 +49,8 @@ class TestCfgToBackend(unittest2.TestCase):
         fnull = open(os.devnull, 'w')
         q = subprocess.Popen(['../alignak_backend_import/cfg_to_backend.py',
                               '--delete',
-                              '--quiet',
-                              'shinken_cfg_files/default/_main.cfg'],
-                             stdout=fnull)
+                              # '--quiet',
+                              'shinken_cfg_files/default/_main.cfg'])
         (_, _) = q.communicate()
         exit_code = q.wait()
         print("Exited with: %d" % exit_code)
@@ -64,8 +66,7 @@ class TestCfgToBackend(unittest2.TestCase):
         q = subprocess.Popen(['../alignak_backend_import/cfg_to_backend.py',
                               '--delete',
                               # '--quiet',
-                              'alignak_cfg_files/alignak-demo/alignak-backend-import.cfg'],
-                             stdout=fnull)
+                              'alignak_cfg_files/alignak-demo/alignak-backend-import.cfg'])
         (_, _) = q.communicate()
         exit_code = q.wait()
         print("Exited with: %d" % exit_code)
@@ -81,9 +82,8 @@ class TestCfgToBackend(unittest2.TestCase):
         fnull = open(os.devnull, 'w')
         q = subprocess.Popen(['../alignak_backend_import/cfg_to_backend.py',
                               '--delete',
-                              '--quiet',
-                              'alignak_cfg_files/alignak_most_recent/alignak-backend-import.cfg'],
-                             stdout=fnull)
+                              # '--quiet',
+                              'alignak_cfg_files/alignak_most_recent/alignak-backend-import.cfg'])
         (_, _) = q.communicate()
         exit_code = q.wait()
         print("Exited with: %d" % exit_code)
