--- test_requirements.txt
@@ -1,21 +1,0 @@
-# Python requirements for unit tests
-unittest2
-# Web test
-webtest
-# Use py.test as test-runner
-pytest
-pytest-cov
-# Let coverage use the most recent version
-coverage
-# Report coverage results to coveralls.io
-coveralls
-# Static code analysis libraries
-pylint
-pycodestyle
-pep257
-# uWSGI server
-uwsgi
-# alignak (most recent develop)
--e git+git://github.com/Alignak-monitoring/alignak.git@develop#egg=alignak
-# alignak_backend (most recent develop)
--e git+git://github.com/Alignak-monitoring-contrib/alignak-backend.git@develop#egg=alignak_backend
